requirements.txt
Flask==3.1.2
requests==2.32.5
python-dotenv==1.1.1

